make rk45 integrator standalone, move to own file
