+++ openscvx/discretization.py
@@ -4,6 +4,7 @@ import numpy as np
 import diffrax  as dfx
 
 from openscvx.config import Config
+from openscvx.integrators import solve_ivp_rk45
 
 SOLVER_MAP = {
     "Tsit5": dfx.Tsit5,
@@ -21,49 +22,17 @@ SOLVER_MAP = {
     "KenCarp5": dfx.KenCarp5
 }
 
+
+
 class RK45_Custom:
     def __init__(self, params):
         self.params = params
-
-    def rk45_step(self, f, t, y, h, *args):
-        def compute_k(f, t, y, h, *args):
-            k1 = f(t, y, *args)
-            k2 = f(t + h/4, y + h*k1/4, *args)
-            k3 = f(t + 3*h/8, y + 3*h*k1/32 + 9*h*k2/32, *args)
-            k4 = f(t + 12*h/13, y + 1932*h*k1/2197 - 7200*h*k2/2197 + 7296*h*k3/2197, *args)
-            k5 = f(t + h, y + 439*h*k1/216 - 8*h*k2 + 3680*h*k3/513 - 845*h*k4/4104, *args)
-            return k1, k2, k3, k4, k5
-        
-        k1, k2, k3, k4, k5 = compute_k(f, t, y, h, *args)
-        y_next = y + h * (25*k1/216 + 1408*k3/2565 + 2197*k4/4104 - k5/5)
-        return y_next
 
     def solve_ivp(self, dVdt, tau_grid, V0, args, method='RK45', t_eval=None):
         if method != 'RK45':
             raise ValueError("Currently, only 'RK45' method is supported.")
         
-        if t_eval is None:
-            t_eval = jnp.linspace(tau_grid[0], tau_grid[1], 50)
-        
-        h = (tau_grid[1] - tau_grid[0]) / (len(t_eval) - 1)
-        V_result = jnp.zeros((len(t_eval), len(V0)))
-        V_result = V_result.at[0].set(V0)
-        
-        if self.params.dev.debug:
-            for i in range(1, len(t_eval)):
-                t = tau_grid[0] + i * h
-                y_next = self.rk45_step(dVdt, t, V_result[i-1], h, *args)
-                V_result = V_result.at[i].set(y_next)
-        else:
-            def body_fun(i, val):
-                t, y, V_result = val
-                y_next = self.rk45_step(dVdt, t, y, h, *args)
-                V_result = V_result.at[i].set(y_next)
-                return (t + h, y_next, V_result)
-
-            _, _, V_result = lax.fori_loop(1, len(t_eval), body_fun, (tau_grid[0], V0, V_result))
-        
-        return V_result
+        return solve_ivp_rk45(dVdt, tau_grid, V0, args, self.params.dev.debug, t_eval)
 
     
 class Diffrax_Prop:
